requirements.txt
django==2.0
mysqlclient
djangorestframework
Pillow
django_rest_swagger